Metrics: Adds (approximate) response duration custom metric.
Adds an custom metric around response duration: the combined time
spent waiting in the (SQS) queue + service time by the worker
process. This provides an end-to-end latency measurement for instances
of work.

Note that this measurement is approximate, since we're measuring the
time difference between clock timestamps on different machines that
cannot be treated as being reliably/sufficiently in-sync at Very Small
time resolutions.

/cc @rclark @jakepruitt @danpat