Note that Subject and Message are parsed
The fact that the Subject and Message are already parsed in the worker's `process.env` caught me a bit off guard today, since generally `process.env` values are strings that need to be coerced into other types. It would be nice to indicate in the docs that this is not the case for these values in the worker's env.